Break out Problems into own component
